+++ workers/ohsome_quality_analyst/utils/helper.py
@@ -30,10 +30,11 @@ def name_to_class(class_type: str, name: str):
     #     - https://julienharbulot.com/python-dynamical-import.html
     class_path = "ohsome_quality_analyst.{0}s.{1}.{2}".format(
         class_type,
-        camel_to_snake(name),
+        hyphen_to_snake(name),
         class_type,
     )
-    return getattr(importlib.import_module(class_path), name)
+    class_name = hyphen_to_camel(name)
+    return getattr(importlib.import_module(class_path), class_name)
 
 
 def camel_to_snake(camel: str) -> str:
@@ -41,17 +42,27 @@ def camel_to_snake(camel: str) -> str:
     return re.sub(r"(?<!^)(?=[A-Z])", "_", camel).lower()
 
 
-def snake_to_lower_camel(snake: str) -> str:
-    """Converts Snake Case to Lower Camel Case"""
+def camel_to_hyphen(camel: str) -> str:
+    """Converts Camel Case to Lower Hyphen Case"""
+    return re.sub(r"(?<!^)(?=[A-Z])", "-", camel).lower()
+
+
+def hyphen_to_camel(hyphen: str):
+    """Convert Lower Hyphen Case to Camel Case"""
+    parts = hyphen.split("-")
+    return "".join(p.title() for p in parts)
+
+
+def hyphen_to_snake(hyphen: str):
+    """Convert Lower Hyphen Case to Camel Case"""
+    parts = hyphen.split("-")
+    return "_".join(parts)
+
+
+def snake_to_hyphen(snake: str) -> str:
+    """Converts Snake Case to Lower Hyphen Case"""
     parts = snake.split("_")
-    return parts[0] + "".join(part.title() for part in parts[1:])
-
-
-def name_to_lower_camel(name: str) -> str:
-    """Convert name to Lower Camel Case"""
-    name = name.replace(" ", "_")
-    name = name.replace("-", "_")
-    return snake_to_lower_camel(name)
+    return "-".join(parts)
 
 
 def get_module_dir(module_name: str) -> str: