+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Current Main
 
+### Breaking Changes
+
+- change API parameter names for POST requests from snake case to lower hyphen ([#398])
+- change indicator and report keys from lower camel case to lower hyphen ([#398])
+- rename API parameter `layerKey` and `layer` to `topic` ([#501])
+- remove JRC related layers and report `jrc_requirements` ([#503])
+
+### New Features
+
+- disable size limit for Mapping Saturation ([#498])
+
 ### Other Changes
 
 - fix pre-commit hooks ([#482])
@@ -9,18 +20,36 @@
 - refactor(db): remove artifacts as well as old init scripts and restructure directories ([#388])
 - feat: disable size limit for Mapping Saturation ([#498])
 - Overhaul docker compose setup ([#483])
-- remove JRC related layers and report `jrc_requirements` ([#503])
+- overhaul docker compose setup ([#483])
+- build(deps): update FastAPI to version 0.92.0 ([#488])
 
 ### How to Upgrade
 
+- rename indicator keys from camel case to lower hyphen ([#398]): E.g. `MappingSaturation` to `mapping-saturation`
+- rename API parameters for POST requests from camel case to lower hyphen ([#398])
+- rename API parameter `layerKey` and `layer` to `topic` ([#501])
 
+| old API parameter | new API parameter |
+| ---               | ---               |
+| `layerKey`        | `topic`           |
+| `includeSvg`      | `include-svg`     |
+| `includeHtml`     | `include-html`    |
+| `featureId`       | `feature-id`      |
+| `fidField`        | `indicators`      |
+>>>>>>> main
 
 [#388]: https://github.com/GIScience/ohsome-quality-analyst/pull/388
+[#398]: https://github.com/GIScience/ohsome-quality-analyst/pull/398
 [#481]: https://github.com/GIScience/ohsome-quality-analyst/pull/481
 [#482]: https://github.com/GIScience/ohsome-quality-analyst/pull/482
 [#483]: https://github.com/GIScience/ohsome-quality-analyst/pull/483
+[#488]: https://github.com/GIScience/ohsome-quality-analyst/pull/488
 [#498]: https://github.com/GIScience/ohsome-quality-analyst/pull/498
+<<<<<<< HEAD
 [#503]: https://github.com/GIScience/ohsome-quality-analyst/pull/503
+=======
+[#501]: https://github.com/GIScience/ohsome-quality-analyst/pull/501
+>>>>>>> main
 
 ## 0.14.0